feat: migrate controller.js to TS
